edycja Portfolio na mapper

+++ backend/TattooMe/src/main/java/TattooMe/TattooMe/service/PortfolioService.java
@@ -3,15 +3,16 @@ package TattooMe.TattooMe.service;
 import TattooMe.TattooMe.dto.portfolio.PortfolioDTO;
 import TattooMe.TattooMe.entity.Portfolio;
 import TattooMe.TattooMe.entity.User;
+import TattooMe.TattooMe.mapper.PortfolioMapper;
 import TattooMe.TattooMe.repository.PortfolioRepository;
 import TattooMe.TattooMe.repository.UserRepository;
 import jakarta.persistence.EntityNotFoundException;
+import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
-import java.util.Base64;
 import java.util.List;
 import java.util.UUID;
 
@@ -19,20 +20,19 @@ import java.util.UUID;
 public class PortfolioService {
     @Autowired
     private PortfolioRepository portfolioRepository;
-
     @Autowired
     private UserRepository userRepository;
+    @Autowired
+    private PortfolioMapper portfolioMapper;
 
     public List<PortfolioDTO> getUserPortfolio(UUID userId) {
-        return portfolioRepository.findAllByUser_Id(userId)
-                .stream()
-                .map(p -> new PortfolioDTO(
-                        p.getId(),
-                        Base64.getEncoder().encodeToString(p.getPicture())
-                ))
-                .toList();
+        userRepository.findById(userId)
+                .orElseThrow(() -> new EntityNotFoundException("Użytkownik nie znaleziony"));
+
+        return portfolioMapper.toDTOList(portfolioRepository.findAllByUser_Id(userId));
     }
 
+    @Transactional
     public PortfolioDTO uploadPortfolioImage(UUID userId, MultipartFile multipartFile) throws IOException {
         User user = userRepository.findById(userId)
                 .orElseThrow(() -> new EntityNotFoundException("Użytkownik nie znaleziony"));
@@ -47,18 +47,18 @@ public class PortfolioService {
         }
 
         Portfolio portfolio = new Portfolio();
-        portfolio.setPicture(multipartFile.getBytes());
         portfolio.setUser(user);
+        portfolio.setPicture(multipartFile.getBytes());
 
         Portfolio saved = portfolioRepository.save(portfolio);
-
-        return new PortfolioDTO(
-                saved.getId(),
-                Base64.getEncoder().encodeToString(saved.getPicture())
-        );
+        return portfolioMapper.toDTO(saved);
     }
 
+    @Transactional
     public void deletePortfolioImage(UUID id) {
+        if (!portfolioRepository.existsById(id)) {
+            throw new EntityNotFoundException("Nie znaleziono zdjęcia portfolio");
+        }
         portfolioRepository.deleteById(id);
     }
 }
